requirements.txt
Pillow>=12.0.0
imageio>=2.37.2
imageio-ffmpeg>=0.6.0